wheel: centralize wheel handling
Move the logic to handle wheels to its own file.
Refactor, no functional changes.
