+++ src/evdev-fallback.h
@@ -254,5 +254,30 @@ fallback_notify_physical_button(struct fallback_dispatch *dispatch,
 				uint64_t time,
 				int button,
 				enum libinput_button_state state);
+void
+fallback_normalize_delta(struct evdev_device *device,
+			 const struct device_coords *delta,
+			 struct normalized_coords *normalized);
+
+void
+fallback_init_wheel(struct fallback_dispatch *dispatch,
+		    struct evdev_device *device);
+
+void
+fallback_wheel_notify_physical_button(struct fallback_dispatch *dispatch,
+				      struct evdev_device *device,
+				      uint64_t time,
+				      int button,
+				      enum libinput_button_state state);
+
+void
+fallback_wheel_process_relative(struct fallback_dispatch *dispatch,
+				struct evdev_device *device,
+				struct input_event *e, uint64_t time);
+
+void
+fallback_wheel_handle_state(struct fallback_dispatch *dispatch,
+			    struct evdev_device *device,
+			    uint64_t time);
 
 #endif
